Refactor and move utility functions into logic #40
