+++ src/controllers/timeline.go
@@ -1,18 +1,13 @@
 package controllers
 
 import (
-	"crypto/md5"
-	"encoding/hex"
+	"github.com/gin-gonic/gin"
+	"github.com/noirbizarre/gonja"
 	"minitwit/src/database"
 	"minitwit/src/functions"
 	"minitwit/src/logic"
 	"minitwit/src/models"
 	"net/http"
-	"strconv"
-	"strings"
-
-	"github.com/gin-gonic/gin"
-	"github.com/noirbizarre/gonja"
 )
 
 func timelineHandlers(router *gin.Engine) {
@@ -33,13 +28,12 @@ func handleUserTimeline(w http.ResponseWriter, r *http.Request, c *gin.Context, 
 
 	request := functions.GetEndpoint(r)
 
-	messages, user, err := logic.GetUserMessages(username)
+	twits, user, err := logic.GetUserTwits(username)
 	if err != nil {
 		http.Error(w, "404 - "+err.Error(), http.StatusNotFound)
 		return
 	}
 
-	twits := ConvertMessagesToTwits(&messages)
 	out, err := timelineTemplate.Execute(gonja.Context{"g": g, "request": request, "messages": twits, "profile_user": user})
 	if err != nil {
 		http.Error(w, err.Error(), http.StatusInternalServerError)
@@ -60,7 +54,7 @@ func handleTimeline(w http.ResponseWriter, r *http.Request, c *gin.Context) {
 
 	user, _ := database.GetUserFromDb(g.User.Username)
 	messages := database.GetUserMessages(g.User.User_id)
-	twits := ConvertMessagesToTwits(&messages)
+	twits := logic.ConvertMessagesToTwits(&messages)
 
 	out, err := timelineTemplate.Execute(gonja.Context{"g": g, "request": request, "messages": twits, "profile_user": user})
 	if err != nil {
@@ -74,30 +68,11 @@ func handlePublicTimeline(w gin.ResponseWriter, r *http.Request, c *gin.Context)
 
 	messages := database.GetAllMessages()
 
-	twits := ConvertMessagesToTwits(&messages)
+	twits := logic.ConvertMessagesToTwits(&messages)
 	//print(string(request))
 	out, err := timelineTemplate.Execute(gonja.Context{"g": g, "request": request, "messages": twits})
 	if err != nil {
 		http.Error(w, err.Error(), http.StatusInternalServerError)
 	}
 	w.Write([]byte(out))
-}
-
-func ConvertMessagesToTwits(messages *[]models.Message) []models.Twit {
-	var twits []models.Twit
-	for _, message := range *messages {
-		twits = append(twits, models.Twit{getGavaterUrl(message.Email, 48), message.Username, strconv.Itoa(int(message.Pubdate)), message.Text})
-	}
-	print(twits)
-	return twits
-
-}
-
-func getGavaterUrl(email string, size int) string {
-	data := []byte(strings.ToLower(strings.TrimSpace(email)))
-	hash := md5.Sum(data)
-	hashStr := hex.EncodeToString(hash[:])
-
-	str := []string{"http://www.gravatar.com/avatar/", hashStr, "?d=identicon&s=", strconv.Itoa(size)}
-	return strings.Join(str, "")
 }
